Make changes so a clean build compiles
`StageDataHolder::getCommonPathPointInfo` and
`StageDataHolder::getCommonPathPointInfoFromRailDataIndex` both had
different return types between their headers and source files. I updated
the header return type based on the source return type because the
source file was more recently updated (see commit b5dc6e8).

CameraUtil.cpp also had a mismatched return type with its header.

`MultiEmitterCallBack.hpp` was deleted for no apparent reason, so I
restored it and gave `JPAEmitterCallback` its own header.
