+++ libs/JSystem/include/JSystem/JParticle/JPAParticleCallBack.hpp
@@ -1,16 +1,8 @@
 #pragma once
 
-class JPABaseEmitter;
-class JPABaseParticle;
+#include <JSystem/JParticle/JPAEmitterCallback.hpp>
 
-class JPAEmitterCallBack {
-    virtual ~JPAEmitterCallBack();
-    virtual void execute(JPABaseEmitter *);
-    virtual void executeAfter(JPABaseEmitter *);
-    virtual void draw(JPABaseEmitter *);
-    virtual void drawAfter(JPABaseEmitter *);
-    virtual void init(JPABaseEmitter *) = 0;
-};
+class JPABaseParticle;
 
 class JPAParticleCallBack {
 public:
