+++ libs/JSystem/include/JSystem/JParticle/MultiEmitterCallBack.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <JSystem/JParticle/JPAEmitterCallBack.hpp>
+
+class MultiEmitterCallBackBase : public JPAEmitterCallBack {
+    virtual ~MultiEmitterCallBackBase();
+    virtual void init(JPABaseEmitter *);
+};
